Deprecate single shortest path path tracking
