+++ src/include/function/gds/gds_object_manager.h
@@ -23,6 +23,9 @@ public:
         nextPosToWrite.fetch_add(1, std::memory_order_relaxed);
         return result;
     }
+    void revertLast() {
+        nextPosToWrite.fetch_sub(1, std::memory_order_relaxed);
+    }
 
     bool hasSpace() const {
         return nextPosToWrite.load(std::memory_order_relaxed) <
