+++ src/include/function/gds/gds_task.h
@@ -2,6 +2,7 @@
 
 #include "common/task_system/task.h"
 #include "function/gds/gds_frontier.h"
+#include "graph/graph.h"
 
 namespace kuzu {
 namespace function {
