+++ src/include/function/gds/rec_joins.h
@@ -2,163 +2,10 @@
 
 #include "function/gds/gds.h"
 #include "function/gds/gds_frontier.h"
+#include "output_writer.h"
 
 namespace kuzu {
-namespace common {
-class ValueVector;
-} // namespace common
-
-namespace graph {
-class Graph;
-} // namespace graph
-
-namespace storage {
-class MemoryBuffer;
-class MemoryManager;
-} // namespace storage
 namespace function {
-class EdgeCompute;
-class FrontierPair;
-class GDSFrontier;
-
-/**
- * A GDSFrontier implementation that keeps the lengths of the paths from a source node to
- * destination nodes. This is a light-weight implementation that can keep lengths up to and
- * including UINT16_MAX - 1. The length stored for the source node is 0. Length UINT16_MAX is
- * reserved for marking nodes that are not visited. The lengths stored per node are effectively the
- * iteration numbers that a node is visited. For example, if the running computation is shortest
- * path, then the length stored is the shortest path length.
- *
- * Note: This class can be used to represent both the current and next frontierPair for shortest
- * paths computations, which have the guarantee that a vertex is part of only one frontier level.
- * Specifically, at iteration i of the shortest path algorithm (iterations start from 0), nodes with
- * mask values i are in the current frontier. Nodes with any other values are not in the frontier.
- * Similarly at iteration i setting a node u active sets its mask value to i+1. To enable this
- * usage, this class contains functions to expose two frontierPair to users, e.g.,
- * getMaskValueFromCur/NextFrontierFixedMask. In the case of shortest path computations, using this
- * class to represent two frontierPair should be faster or take less space than keeping two separate
- * frontierPair.
- *
- * However, this is not necessary and the caller can also use this to represent a single frontier.
- */
-class PathLengths : public GDSFrontier {
-    friend class SinglePathLengthsFrontierPair;
-
-public:
-    static constexpr uint16_t UNVISITED = UINT16_MAX;
-
-    explicit PathLengths(std::unordered_map<common::table_id_t, uint64_t> nodeTableIDAndNumNodes,
-        storage::MemoryManager* mm);
-
-    uint16_t getMaskValueFromCurFrontierFixedMask(common::offset_t nodeOffset) {
-        return getCurFrontierFixedMask()[nodeOffset].load(std::memory_order_relaxed);
-    }
-
-    uint16_t getMaskValueFromNextFrontierFixedMask(common::offset_t nodeOffset) {
-        return getNextFrontierFixedMask()[nodeOffset].load(std::memory_order_relaxed);
-    }
-
-    inline bool isActive(common::nodeID_t nodeID) override {
-        return getCurFrontierFixedMask()[nodeID.offset] ==
-               curIter.load(std::memory_order_relaxed) - 1;
-    }
-
-    inline void setActive(common::nodeID_t nodeID) override {
-        getNextFrontierFixedMask()[nodeID.offset].store(curIter.load(std::memory_order_relaxed),
-            std::memory_order_relaxed);
-    }
-
-    void incrementCurIter() { curIter.fetch_add(1, std::memory_order_relaxed); }
-
-    void fixCurFrontierNodeTable(common::table_id_t tableID);
-
-    void fixNextFrontierNodeTable(common::table_id_t tableID);
-
-    uint64_t getNumNodesInCurFrontierFixedNodeTable() {
-        KU_ASSERT(curFrontierFixedMask.load(std::memory_order_relaxed) != nullptr);
-        return maxNodesInCurFrontierFixedMask.load(std::memory_order_relaxed);
-    }
-
-    uint16_t getCurIter() { return curIter.load(std::memory_order_relaxed); }
-
-private:
-    std::atomic<uint16_t>* getCurFrontierFixedMask() {
-        auto retVal = curFrontierFixedMask.load(std::memory_order_relaxed);
-        KU_ASSERT(retVal != nullptr);
-        return retVal;
-    }
-
-    std::atomic<uint16_t>* getNextFrontierFixedMask() {
-        auto retVal = nextFrontierFixedMask.load(std::memory_order_relaxed);
-        KU_ASSERT(retVal != nullptr);
-        return retVal;
-    }
-
-private:
-    // We do not need to make nodeTableIDAndNumNodesMap and masks atomic because they should only
-    // be accessed by functions that are called by the "master GDS thread" (so not accessed inside
-    // the parallel functions in GDSUtils, which are called by other "worker threads").
-    std::unordered_map<common::table_id_t, uint64_t> nodeTableIDAndNumNodesMap;
-    common::table_id_map_t<std::unique_ptr<storage::MemoryBuffer>> masks;
-    // See FrontierPair::curIter. We keep a copy of curIter here because PathLengths stores
-    // iteration numbers for vertices and uses them to identify which vertex is in the frontier.
-    std::atomic<uint16_t> curIter;
-    std::atomic<common::table_id_t> curTableID;
-    std::atomic<uint64_t> maxNodesInCurFrontierFixedMask;
-    std::atomic<std::atomic<uint16_t>*> curFrontierFixedMask;
-    std::atomic<std::atomic<uint16_t>*> nextFrontierFixedMask;
-};
-
-class SinglePathLengthsFrontierPair : public FrontierPair {
-    friend struct AllSPLengthsEdgeCompute;
-    friend struct AllSPPathsEdgeCompute;
-    friend struct SingleSPLengthsEdgeCompute;
-    friend struct SingleSPPathsEdgeCompute;
-
-public:
-    explicit SinglePathLengthsFrontierPair(std::shared_ptr<PathLengths> pathLengths,
-        uint64_t maxThreadsForExec)
-        : FrontierPair(pathLengths /* curFrontier */, pathLengths /* nextFrontier */,
-              1 /* initial num active nodes */, maxThreadsForExec),
-          pathLengths{pathLengths}, morselDispatcher(maxThreadsForExec) {}
-
-    bool getNextRangeMorsel(FrontierMorsel& frontierMorsel) override;
-
-    void initRJFromSource(common::nodeID_t source) override;
-
-    void beginFrontierComputeBetweenTables(common::table_id_t curFrontierTableID,
-        common::table_id_t nextFrontierTableID) override;
-
-    void beginNewIterationInternalNoLock() override { pathLengths->incrementCurIter(); }
-
-private:
-    std::shared_ptr<PathLengths> pathLengths;
-    FrontierMorselDispatcher morselDispatcher;
-};
-
-class DoublePathLengthsFrontierPair : public FrontierPair {
-    friend struct VarLenJoinsEdgeCompute;
-
-public:
-    explicit DoublePathLengthsFrontierPair(
-        std::unordered_map<common::table_id_t, uint64_t> nodeTableIDAndNumNodes,
-        uint64_t maxThreadsForExec, storage::MemoryManager* mm);
-
-    bool getNextRangeMorsel(FrontierMorsel& frontierMorsel) override;
-
-    void initRJFromSource(common::nodeID_t source) override;
-
-    void beginFrontierComputeBetweenTables(common::table_id_t curFrontierTableID,
-        common::table_id_t nextFrontierTableID) override;
-
-    void beginNewIterationInternalNoLock() override {
-        curFrontier->ptrCast<PathLengths>()->incrementCurIter();
-        nextFrontier->ptrCast<PathLengths>()->incrementCurIter();
-    }
-
-private:
-    std::unique_ptr<FrontierMorselDispatcher> morselDispatcher;
-};
 
 struct RJBindData final : public function::GDSBindData {
     static constexpr uint16_t DEFAULT_MAXIMUM_ALLOWED_UPPER_BOUND = (uint16_t)255;
@@ -195,80 +42,6 @@ struct RJBindData final : public function::GDSBindData {
     }
 };
 
-struct RJOutputs {
-public:
-    explicit RJOutputs(common::nodeID_t sourceNodeID);
-    virtual ~RJOutputs() = default;
-
-    virtual void initRJFromSource(common::nodeID_t) {}
-    virtual void beginFrontierComputeBetweenTables(common::table_id_t curFrontierTableID,
-        common::table_id_t nextFrontierTableID) = 0;
-    // This function is called after the recursive join computation stage, at the stage when the
-    // outputs that are stored in RJOutputs is being written to FactorizedTable.
-    virtual void beginWritingOutputsForDstNodesInTable(common::table_id_t tableID) = 0;
-    template<class TARGET>
-    TARGET* ptrCast() {
-        return common::ku_dynamic_cast<TARGET*>(this);
-    }
-
-public:
-    common::nodeID_t sourceNodeID;
-};
-
-struct SPOutputs : public RJOutputs {
-public:
-    explicit SPOutputs(std::unordered_map<common::table_id_t, uint64_t> nodeTableIDAndNumNodes,
-        common::nodeID_t sourceNodeID, storage::MemoryManager* mm);
-
-public:
-    std::shared_ptr<PathLengths> pathLengths;
-};
-/**
- * Helper class to write paths to a ValueVector. The ValueVector should be a ListVector. The path
- * is written in reverse order, so the calls to addNewNodeID should be done in the reverse order of
- * the nodes that should be placed into the ListVector.
- */
-class PathVectorWriter {
-public:
-    // This is a temporary solution to provide backward compatibility for single shortest path.
-    // TODO(Xiyang): remove this once we track properly for single shortest path.
-    explicit PathVectorWriter(common::ValueVector* nodeIDsVector)
-        : nodeIDsVector{nodeIDsVector}, edgeIDsVector{nullptr} {}
-    PathVectorWriter(common::ValueVector* nodeIDsVector, common::ValueVector* edgeIDsVector)
-        : nodeIDsVector{nodeIDsVector}, edgeIDsVector{edgeIDsVector} {}
-    void beginWritingNewPath(uint64_t length) const;
-    void addNode(common::nodeID_t nodeID, common::sel_t pos) const;
-    void addEdge(common::relID_t edgeID, common::sel_t pos) const;
-    void addNodeEdge(common::nodeID_t nodeID, common::relID_t edgeID, common::sel_t pos) const;
-
-public:
-    common::ValueVector* nodeIDsVector;
-    common::ValueVector* edgeIDsVector;
-};
-
-class RJOutputWriter {
-public:
-    explicit RJOutputWriter(main::ClientContext* context, RJOutputs* rjOutputs);
-    virtual ~RJOutputWriter() = default;
-
-    void beginWritingForDstNodesInTable(common::table_id_t tableID) {
-        rjOutputs->beginWritingOutputsForDstNodesInTable(tableID);
-    }
-
-    virtual bool skipWriting(common::nodeID_t dstNodeID) const = 0;
-
-    virtual void write(processor::FactorizedTable& fTable, common::nodeID_t dstNodeID) const = 0;
-
-    virtual std::unique_ptr<RJOutputWriter> copy() = 0;
-
-protected:
-    main::ClientContext* context;
-    RJOutputs* rjOutputs;
-    std::unique_ptr<common::ValueVector> srcNodeIDVector;
-    std::unique_ptr<common::ValueVector> dstNodeIDVector;
-    std::vector<common::ValueVector*> vectors;
-};
-
 // Wrapper around the data that needs to be stored during the computation of a recursive joins
 // computation from one source. Also contains several initialization functions.
 struct RJCompState {
@@ -300,29 +73,6 @@ struct RJCompState {
         frontierPair->beginFrontierComputeBetweenTables(curFrontierTableID, nextFrontierTableID);
         outputs->beginFrontierComputeBetweenTables(curFrontierTableID, nextFrontierTableID);
     }
-};
-
-class SPOutputWriterDsts : public RJOutputWriter {
-public:
-    explicit SPOutputWriterDsts(main::ClientContext* context, RJOutputs* rjOutputs)
-        : RJOutputWriter(context, rjOutputs){};
-
-    void write(processor::FactorizedTable& fTable, common::nodeID_t dstNodeID) const override;
-
-    bool skipWriting(common::nodeID_t dstNodeID) const override {
-        return dstNodeID == rjOutputs->ptrCast<SPOutputs>()->sourceNodeID ||
-               PathLengths::UNVISITED ==
-                   rjOutputs->ptrCast<SPOutputs>()
-                       ->pathLengths->getMaskValueFromCurFrontierFixedMask(dstNodeID.offset);
-    }
-
-    std::unique_ptr<RJOutputWriter> copy() override {
-        return std::make_unique<SPOutputWriterDsts>(context, rjOutputs);
-    }
-
-protected:
-    virtual void writeMoreAndAppend(processor::FactorizedTable& fTable, common::nodeID_t dstNodeID,
-        uint16_t length) const;
 };
 
 class RJAlgorithm : public GDSAlgorithm {
