+++ test/test_files/function/gds/rec_joins_small.test
@@ -405,24 +405,24 @@ Runtime exception: Shortest path operations only works for positive upper bound 
 -STATEMENT PROJECT GRAPH PK (person1, knows11)
            MATCH (a:person1) WHERE a.ID = 0
            CALL single_sp_paths(PK, a, 5, true)
-           RETURN a.ID, _node.ID, length, pathNodeIDs;
+           RETURN a.ID, _node.ID, length, pathNodeIDs, pathEdgeIDs;
 ---- 4
-0|1|1|[]
-0|2|2|[0:1]
-0|3|3|[0:1,0:2]
-0|4|3|[0:1,0:2]
+0|1|1|[]|[2:0]
+0|2|2|[0:1]|[2:0,2:1]
+0|3|3|[0:1,0:2]|[2:0,2:1,2:2]
+0|4|3|[0:1,0:2]|[2:0,2:1,2:4]
 
 -LOG SingleSPPathsMultilabel1
 -STATEMENT PROJECT GRAPH PK (person1, person2, knows12, knows21)
            MATCH (a:person1) WHERE a.ID = 0
            CALL single_sp_paths(PK, a, 30, true)
-           RETURN a.ID, _node.ID, length, pathNodeIDs;
+           RETURN a.ID, _node.ID, length, pathNodeIDs, pathEdgeIDs;
 ---- 5
-0|1|1|[]
-0|2|2|[1:1]
-0|3|3|[1:1,0:2]
-0|4|3|[1:1,0:2]
-0|4|4|[1:1,0:2,1:3]
+0|1|1|[]|[3:0]
+0|2|2|[1:1]|[3:0,4:1]
+0|3|3|[1:1,0:2]|[3:0,4:1,3:2]
+0|4|3|[1:1,0:2]|[3:0,4:1,3:4]
+0|4|4|[1:1,0:2,1:3]|[3:0,4:1,3:2,4:3]
 
 -LOG SingleSPPathsMultilabel2
 -STATEMENT PROJECT GRAPH PK (person1, person2, knows11, knows12, knows21, knows22)
